# Conflicts:
#	libs/ledger/include/ledger/consensus/consensus.hpp
#	libs/ledger/src/consensus/consensus.cpp
#	libs/ledger/src/consensus/stake_manager.cpp

+++ libs/ledger/include/ledger/consensus/consensus.hpp
@@ -50,7 +50,7 @@ public:
 
   Consensus(StakeManagerPtr stake, BeaconServicePtr beacon, MainChain const &chain,
             StorageInterface &storage, Identity mining_identity, uint64_t aeon_period,
-            uint64_t max_cabinet_size, uint32_t block_interval_ms = 1000);
+            uint64_t max_cabinet_size, uint64_t block_interval_ms = 1000);
 
   void         UpdateCurrentBlock(Block const &current) override;
   NextBlockPtr GenerateNextBlock() override;
@@ -65,7 +65,6 @@ public:
 
 private:
   static constexpr std::size_t HISTORY_LENGTH = 1000;
-  //  static constexpr uint64_t    INVALID_BLOCK  = 0xFFFFFFFFFFFFFFFFllu;
 
   using Cabinet        = StakeManager::Cabinet;
   using CabinetPtr     = std::shared_ptr<Cabinet const>;
@@ -92,7 +91,7 @@ private:
 
   uint64_t       default_start_time_ = 0;
   CabinetHistory cabinet_history_{};  ///< Cache of historical cabinets
-  uint32_t       block_interval_ms_{std::numeric_limits<uint32_t>::max()};
+  uint64_t       block_interval_ms_{std::numeric_limits<uint64_t>::max()};
 
   CabinetPtr GetCabinet(Block const &previous);
   bool       ValidMinerForBlock(Block const &previous, chain::Address const &address);
